requirements.txt
peewee==3.13.2
Flask==1.1.1
pycrypto==2.6.1
requests==2.23.0
python-dateutil==2.8.1

